RENAME VALUATION PAGE
- Renaming valuation page to metaverseexplorer
