+++ components/Heatmap/MapSearch.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react'
 import { AiFillQuestionCircle, AiOutlineSearch } from 'react-icons/ai'
 import { ValuationTile } from '../../lib/heatmap/heatmapCommonTypes'
 import { getState, typedKeys } from '../../lib/utilities'
-import { ValuationState } from '../../pages/valuation'
+import { ValuationState } from '../../pages/metaverseexplorer'
 import SearchLandButton from './SearchLandButton'
 import Image from 'next/image'
 import { Tooltip } from '@mui/material'
